+++ Tests/NetworkLayerTests/Classes/Tests/IntegrationTests/RequestProcessorAuthenticationTests.swift
@@ -74,7 +74,7 @@ final class RequestProcessorAuthenicationTests: XCTestCase {
     // MARK: Private
 
     private func test_failAuthentication(adaptError: Error?, refreshError: Error?, expectedError: Error) async throws {
-        class FailInterceptor: IAuthenticationInterceptor {
+        class FailInterceptor: IAuthenticationInterceptor, @unchecked Sendable {
             let adaptError: Error?
             let refreshError: Error?
 
@@ -125,7 +125,7 @@ final class RequestProcessorAuthenicationTests: XCTestCase {
 
 // MARK: - AuthInterceptor
 
-private final class AuthInterceptor: IAuthenticationInterceptor {
+private final class AuthInterceptor: IAuthenticationInterceptor, @unchecked Sendable {
     var token: Token!
 
     private var attempts = 0