+++ CHANGELOG.md
@@ -2,7 +2,21 @@
 All notable changes to this project will be documented in this file.
 
 #### 1.x Releases
+- `1.1.x` Release Candidates - [1.1.0-rc.1](#110-rc1)
 - `1.0.x` Releases - [1.0.0](#100)
+
+## [1.1.0-rc.1](https://github.com/space-code/network-layer/releases/tag/1.1.0-rc.1)
+Released on 2024-12-25.
+
+#### Added
+- Bump the Swift version to 6.0.
+  - Added in Pull Request [#3](https://github.com/space-code/network-layer/pull/3).
+
+#### Fixed
+- Fix the `inout` parameter in the `IAuthenticator` protocol
+  - Fixed in Pull Request [#2](https://github.com/space-code/network-layer/pull/2).
+- Fix the package builiding.
+  - Fixed in Pull Request [#6](https://github.com/space-code/network-layer/pull-6).
 
 ## [1.0.0](https://github.com/space-code/network-layer/releases/tag/1.0.0)
 Released on 2023-12-04.